Move `HtmlWriter` into a private package.
